Lesson 20 AWS

+++ app/src/main/resources/db/migration/V6__Create_invoice_entry.sql
@@ -1,12 +1,12 @@
 CREATE TABLE public.invoice_entry
 (
-    id                     bigserial       NOT NULL,
+    id                     integer       NOT NULL,
     description            character varying(50),
-    quantity               numeric(10, 2) NOT NULL DEFAULT 1,
+    quantity               integer NOT NULL DEFAULT 1,
     net_price              numeric(10, 2) NOT NULL,
     vat_value              numeric(10, 2) NOT NULL,
-    vat_rate               bigint         NOT NULL,
-    expense_related_to_car bigint,
+    vat_rate               bigserial         NOT NULL,
+    expense_related_to_car integer,
     PRIMARY KEY (id)
 );
