+++ app/src/main/resources/db/migration/V7__Connect_invoice_with_invoice_entry.sql
@@ -1,7 +1,7 @@
 CREATE TABLE public.invoice_invoice_entry
 (
-    invoice_id       bigint NOT NULL,
-    invoice_entry_id bigint NOT NULL,
+    invoice_id       integer NOT NULL,
+    invoice_entry_id integer NOT NULL,
     PRIMARY KEY (invoice_id, invoice_entry_id)
 );
